Login creation and data posting
The assumption is that you have a database created and the php access'
the database then goes through the process of creating a login account
in the database. By using the login you then create entries into the
table based on the given input forms. Once a submit has been made the
data is then matched with the unique id of the user and posted to said
database and displayed afterwards on the front end.